Expanded the landing page
I think the landing page is a bit too minimal. It would benefit from a slightly more elaborate introduction for the uninitiated, and from having direct links to specific information that users are likely looking for. 

The idea is to make things a bit easier to find for both newcomers, and for people looking for something specific, like the source code or example Jupyter notebooks.

Before, `velocyto.py` linked to the python documentation, `velocyto.R` linked to the source code (I assume R does not have the same style of documentating itself as Python does). Neither link signified where it would send you. Example of why this is annoying: I was looking for the python source code, searched for "velocyto", landed on this page, clicked through to the documentation (which I did not expect), then had to look through the documentation to find a link to the repository.

Links and explanations were added to both implementations. I also inlined the entire velocyto.R documentation, since most users just want to install it quickly, and having to navigate to the github page for documentation that is so short is pretty silly.

Also: yes, I know that I link to the github repo in both implementation sections and the note about this being alpha software. Easy-to-use documentation should contain relevant information in places where users *are most likely to look*, even if this creates some redundancy.